fix guided tour menu item for active perspective
